pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aakash-airflow-snowflake-dag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow-providers-snowflake>=6.5.4",
    "snowflake-connector-python>=3.17.4",
]

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]  # look for packages in the root directory
exclude = [
    "dags*",
    "logs*", 
    "config*",
    "plugins*",
    "screenshots*",
    "tests*",
    ".git*",
    ".venv*"
]
